% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentspace.R
\name{latent.netdep}
\alias{latent.netdep}
\title{Generate latent variable dependent network}
\usage{
latent.netdep(n.node, rho)
}
\arguments{
\item{n.node}{The number of nodes in network}

\item{rho}{correlation coefficient}
}
\description{
Generate latent variable dependent network
}
\examples{



}
