% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettest.R
\name{MoranI}
\alias{MoranI}
\title{Moran's I statistic}
\usage{
MoranI(A, Y)
}
\arguments{
\item{A}{\code{[n x n]} adjacency matrix or general relational weight matrix of which \eqn{A_ij} indicates relationship from \eqn{i} to \eqn{j}.}

\item{Y}{a vector of \eqn{n} continuous or binary, one-dimensional observations.}
}
\value{
\item{\code{moran}}{a standardized Moran's \eqn{I} statistic.}
}
\description{
This function calculates Moran's I statistic given adjacency matrix and outcome.
The statistic can be used for testing network dependence.
}
\author{
Youjin Lee
}
