% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettest.R
\name{phi.stat}
\alias{phi.stat}
\title{Standardized \eqn{\Phi} statistic}
\usage{
phi.stat(A, Y)
}
\arguments{
\item{A}{\code{[n x n]} adjacency matrix or general relational weight matrix of which \eqn{A_ij} indicates relationship from \eqn{i} to \eqn{j}.}

\item{Y}{a vector of \eqn{n} nominal or binary, one-dimensional observations.}
}
\value{
\item{\code{phi}}{a standardized \eqn{\Phi} statistic.}
}
\description{
A test statistic of \eqn{\Phi} is to test network dependence in categorical observations.
}
\author{
Youjin Lee
}
